#231 [mod]: 알림 띄우기 로직 분리
