+++ app/src/main/java/com/byeboo/app/ByeBoo.kt
@@ -7,6 +7,7 @@ import android.os.Build
 import androidx.appcompat.app.AppCompatDelegate
 import com.byeboo.app.core.util.MixpanelUtil
 import com.byeboo.app.fcm.ByebooMessagingService
+import com.byeboo.app.fcm.ByebooNotificationHandler
 import com.kakao.sdk.common.KakaoSdk
 import dagger.hilt.android.HiltAndroidApp
 import javax.inject.Inject
@@ -47,8 +48,8 @@ class ByeBoo : Application() {
     private fun initNotificationChannel() {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
             val channel = NotificationChannel(
-                ByebooMessagingService.CHANNEL_ID,
-                ByebooMessagingService.CHANNEL_NAME,
+                ByebooNotificationHandler.CHANNEL_ID,
+                ByebooNotificationHandler.CHANNEL_NAME,
                 NotificationManager.IMPORTANCE_HIGH
             )
             val notificationManager = getSystemService(NotificationManager::class.java)
