+++ app/src/main/java/com/byeboo/app/fcm/ByebooNotificationHandler.kt
@@ -1,0 +1,54 @@
+package com.byeboo.app.fcm
+
+import android.app.NotificationManager
+import android.app.PendingIntent
+import android.content.Context
+import android.content.Context.NOTIFICATION_SERVICE
+import android.content.Intent
+import androidx.core.app.NotificationCompat
+import androidx.core.app.NotificationManagerCompat
+import com.byeboo.app.R
+import com.byeboo.app.presentation.main.MainActivity
+import dagger.hilt.android.qualifiers.ApplicationContext
+import javax.inject.Inject
+
+class ByebooNotificationHandler @Inject constructor(
+    @ApplicationContext private val context: Context
+) {
+    fun showNotification(
+        title: String?,
+        message: String?,
+        questId: String
+    ) {
+        val notifyId = System.currentTimeMillis().toInt()
+
+        val intent = Intent(context, MainActivity::class.java).apply {
+            putExtra("questId", questId)
+            addFlags(Intent.FLAG_ACTIVITY_SINGLE_TOP)
+        }
+
+        val pendingIntent = PendingIntent.getActivity(
+            context,
+            notifyId,
+            intent,
+            PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
+        )
+
+        val builder = NotificationCompat.Builder(context, CHANNEL_ID)
+            .setSmallIcon(R.mipmap.ic_launcher)
+            .setContentTitle(title)
+            .setContentText(message)
+            .setPriority(NotificationManagerCompat.IMPORTANCE_HIGH)
+            .setDefaults(NotificationCompat.DEFAULT_ALL)
+            .setAutoCancel(true)
+            .setContentIntent(pendingIntent)
+
+        val notificationManager = context.getSystemService(NOTIFICATION_SERVICE) as NotificationManager
+        notificationManager.notify(notifyId, builder.build())
+    }
+
+    companion object {
+        const val CHANNEL_ID = "BYEBOO"
+        const val CHANNEL_NAME = "Byeboo 알림 채널"
+    }
+}
